v1.2.0
- Rebuild with gcc-min
- Switch to nsp, codecov
- Update test scripts for cover all versions
- Update module index package.json
