--- index.js
@@ -1,8 +1,0 @@
-/**
- * Load core
- * @ndaidong
-**/
-
-var main = require('./dist/scheduler');
-main.version = require('./package').version;
-module.exports = main;
