--- test/config.js
@@ -1,6 +1,0 @@
-
-var scheduler = require('../dist/scheduler.min');
-
-module.exports = {
-  scheduler
-};
