--- test/specs/scheduler.js
@@ -1,139 +1,0 @@
-/**
- * Testing
- * @ndaidong
- */
-
-var test = require('tape');
-var sinon = require('sinon');
-
-var config = require('../config');
-var scheduler = config.scheduler;
-
-test('Testing scheduler.every(String pattern, Function callback) method:', (assert) => {
-
-  let clock = sinon.useFakeTimers();
-  let callback = sinon.spy();
-
-  scheduler.every('5s', callback);
-
-  clock.tick(21000);
-  assert.deepEquals(callback.callCount, 4, 'Callback must be called 4 times');
-
-  clock.restore();
-  assert.end();
-});
-
-test('Testing scheduler.once(String pattern, Function callback) method:', (assert) => {
-
-  let clock = sinon.useFakeTimers();
-  let callback = sinon.spy();
-
-  scheduler.once('5s', callback);
-
-  clock.tick(21000);
-  assert.deepEquals(callback.callCount, 1, 'Callback must be called 1 time');
-
-  clock.restore();
-  assert.end();
-});
-
-test('Testing if called on Wednesday:', (assert) => {
-
-  let callback = sinon.spy();
-  let t = new Date(2016, 3, 18, 14, 0, 0);
-
-  // set start time is Monday, 14:00:00 04/18/2016
-  let clock = sinon.useFakeTimers(t.getTime());
-
-  scheduler.once('wed 15', callback);
-
-  clock.tick(6e4 * 60 * 24 * 3);
-  assert.deepEquals(callback.callCount, 1, 'Callback must be called 1 time');
-
-  clock.restore();
-  assert.end();
-});
-
-
-test('Testing other methods:', (assert) => {
-
-  let t = new Date(2016, 3, 18, 14, 0, 0);
-
-  // set start time is 14:00:00 04/18/2016
-  let clock = sinon.useFakeTimers(t.getTime());
-  let hourly = sinon.spy();
-  let daily = sinon.spy();
-  let monthly = sinon.spy();
-  let yearly = sinon.spy();
-
-  // every hour at 15th min
-  scheduler.hourly('15 00', hourly);
-
-  // every day at 14:15:00
-  scheduler.daily('14 15 00', daily);
-
-  // every month at 7th 14:15:00
-  scheduler.monthly('07 14 15 00', monthly);
-
-  // every year at May 12 14:15:00
-  scheduler.yearly('05 12 14 15 00', yearly);
-
-  // go ahead 2 hours --> 16:00:00 04/18/2016
-  // events at 14:15 and 15:15
-  clock.tick(6e4 * 60 * 2);
-  assert.deepEquals(hourly.callCount, 2, 'Hourly task must be called 2 times');
-
-  // go ahead 2 days --> 16:00:00 04/20/2016
-  // events at
-  // - 14:15 04/18/2016
-  // - 14:15 04/19/2016
-  // - 14:15 04/20/2016
-  clock.tick(6e4 * 60 * 24 * 2);
-  assert.deepEquals(daily.callCount, 3, 'Daily task must be called 3 times');
-
-  // go ahead 120 days --> 16:00:00 08/18/2016
-  // events at:
-  // - 14:15 05/7/2016
-  // - 14:15 06/7/2016
-  // - 14:15 07/7/2016
-  // - 14:15 08/7/2016
-  clock.tick(6e4 * 60 * 24 * 120);
-  assert.deepEquals(monthly.callCount, 4, 'Monthly task must be called 4 times');
-
-  // go ahead 1500 days --> 16:00:00 05/29/2020
-  // events at:
-  // - 14:15 05/12/2016
-  // - 14:15 05/12/2017
-  // - 14:15 05/12/2018
-  // - 14:15 05/12/2019
-  // - 14:15 05/12/2020
-  clock.tick(6e4 * 60 * 24 * 1500);
-  assert.deepEquals(yearly.callCount, 5, 'Yearly task must be called 5 times');
-
-  clock.restore();
-  assert.end();
-});
-
-
-test('Testing .unregister() method:', (assert) => {
-
-  let callback = sinon.spy();
-  let t = new Date(2016, 3, 18, 14, 0, 0);
-
-  // set start time is Monday, 14:00:00 04/18/2016
-  let clock = sinon.useFakeTimers(t.getTime());
-
-  let tid = scheduler.once('wed 15', callback);
-  assert.ok(tid, 'It must return an ID');
-
-  if (tid) {
-    let re = scheduler.unregister(tid);
-    assert.ok(re === true, `unregister(${tid}) must be done`);
-  }
-
-  clock.tick(6e4 * 60 * 24 * 3);
-  assert.deepEquals(callback.callCount, 0, 'Callback must be called 0 time');
-
-  clock.restore();
-  assert.end();
-});
